API reference

This is the official API reference of DataWeaver. Note that it can also be queried interactively from the Julia REPL using the help mode:

julia> using DataWeaver
julia>?
help?> DataWeaver

CellArray type and basic constructors

Missing docstring.

Missing docstring for CellArray. Check Documenter's build log for details.

Convenience type aliases and constructors

Index

DataWeaver.CPUCellArray

DataWeaver.CuCellArray

DataWeaver.ROCCellArray

Documentation

Functions – additional to standard AbstractArray functionality

Index

DataWeaver.adios2_config

DataWeaver.adios2_init

DataWeaver.finalize_adios

DataWeaver.read_setup

DataWeaver.write

DataWeaver.write_setup

Documentation

DataWeaver.adios2_config — Function

If the ADIOS2 config file is not provided, this function uses the XML template to create the file automatically with parameters provided by the user. Returns the XML path to the new file.

DataWeaver.adios2_init — Method

Intialize ADIOS2. Supports several configuration options: MPI, serial, existing XML config file or a new file with specified parameters.

DataWeaver.finalize_adios — Method

Close the ADIOS2 engine once all steps have been performed.

DataWeaver.read_setup — Function

Initialize io in read mode and the corresponding engine for reading data.

DataWeaver.write — Function

Perform the update using specified variables.

DataWeaver.write_setup — Method

Initialize io in write mode and the corresponding engine for writing data.